web-cmd: fix json output by finishing request
With Tornado, when the get() function exits, the request is not
finished. This lets one make an asynchronous request and close the
request manually in the callback.

The requests that are made with ajax currently don't close the
connexions. Add a call to self.finish() after having output the json
information.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>